requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
redis[asyncio]==5.0.8
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
httpx==0.27.2
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
